Move security activities into security package
