+++ app/src/main/scala/com/waz/zclient/security/SecureActivity.scala
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.waz.zclient
+package com.waz.zclient.security
 
 import android.app.Activity
 import android.content.Intent
