+++ app/src/main/scala/com/waz/zclient/security/SecurityCheckActivity.scala
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.waz.zclient
+package com.waz.zclient.security
 
 import android.app.Activity
 import android.content.Context
@@ -23,9 +23,9 @@ import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import com.waz.log.BasicLogging.LogTag.DerivedLogTag
 import com.waz.threading.Threading.Implicits.Background
+import com.waz.zclient.BuildConfig
 import com.waz.zclient.log.LogUI._
 import com.waz.zclient.security.SecurityCheckList.{Action, Check}
-import com.waz.zclient.security.{BlockWithDialog, RootDetectionCheck, SecurityCheckList}
 
 import scala.collection.mutable.ListBuffer
